separated AddChild from ManageChild component
