+++ src/screens/ManageChildren/index.tsx
@@ -1,24 +1,28 @@
-
-import { Box, Button, Center, Column, Heading, Row, Text, Input, VStack, FormControl, HStack } from 'native-base';
-import React, { useState, useEffect } from 'react';
-import { Controller, useForm } from 'react-hook-form';
+import {
+  Box,
+  Button,
+  Center,
+  Column,
+  Heading,
+  Row,
+  VStack,
+} from 'native-base';
+import React, { useState } from 'react';
+import {  useForm } from 'react-hook-form';
 import EditIcon from '../../appIcons/Edit';
-import { CreateChildInput, useCreateChildMutation } from '../../generated/graphql';
+import {
+  CreateChildInput,
+  useCreateChildMutation,
+} from '../../generated/graphql';
 // import EditChildSheet from './EditChildSheet';
-import * as SecureStore from 'expo-secure-store';
+import AddChild from './AddChild';
 
 interface Props {}
 
-interface ChildData {
-  name: string;
-  dateOfBirth: string;
-}
 
 
+function ManageChildren({}: Props) {
 
-
-function ManageChildren({ }: Props) {
-  const { handleSubmit, control, formState: { errors } } = useForm<ChildData>();
   const { loading: userLoading, data: userData } = {
     loading: false,
     data: { getUser: { children: [] } },
@@ -28,100 +32,12 @@ function ManageChildren({ }: Props) {
   const [selectedChild, setSelectedChild] = useState<any>(); // TODO: add type
   const [showAddChildForm, setShowAddChildForm] = useState<boolean>(false);
   const children = userData?.getUser?.children;
-  
 
-
-  const [createChildMutation, { data, loading, error }] = useCreateChildMutation({
-    async onCompleted(data) {
-      if (!data?.createChild) {
-        setErrorMessage('Error logging in, check credentials and try again.');
-        console.log('error details', data);
-      } else {
-        setErrorMessage('');
-        console.log('success', data);
-      }
-    },
-  });
-  
-
-
-  const onSubmit = async (data: CreateChildInput) => {
-    console.log(data, 'data', userData, 'userData')
-    await createChildMutation({
-      variables: {
-        input: {
-          dateOfBirth: data.dateOfBirth,
-          name: data.name,
-        }
-      },
-    });
+  const toggleAddChildForm = () => {
+    setShowAddChildForm(!showAddChildForm);
   };
 
-  const renderAddChildForm = () => {
-    if (!showAddChildForm) {
-      return null;
-    }
-
-    return (
-      <VStack space={3} mt="5">
-        <FormControl>
-            <FormControl.Label>Name</FormControl.Label>
-            <Controller
-              control={control}
-              name="name"
-              render={({ field: { onChange, onBlur, value } }) => (
-                <Input
-                  autoCapitalize="none"
-                  onChangeText={onChange}
-                  onBlur={onBlur}
-                  value={value}
-                />
-              )}
-            />
-            {errors?.name && (
-              <FormControl.ErrorMessage _text={{ fontSize: 'xs' }}>
-                {errors?.name?.message}
-              </FormControl.ErrorMessage>
-            )}
-          </FormControl>
-          <FormControl>
-            <FormControl.Label>Date of Birth</FormControl.Label>
-            <Controller
-              control={control}
-              name="dateOfBirth"
-              render={({ field: { onChange, onBlur, value } }) => (
-                <Input
-                  onChangeText={onChange}
-                  onBlur={onBlur}
-                  value={value}
-                />
-              )}
-              
-            />
-            {errors?.dateOfBirth && (
-              <FormControl.ErrorMessage>
-                {errors?.dateOfBirth?.message}
-              </FormControl.ErrorMessage>
-            )}
-          </FormControl>
-          {errorMessage ? (
-            <FormControl.ErrorMessage _text={{ fontSize: 'xs' }}>
-              {errorMessage}
-            </FormControl.ErrorMessage>
-          ) : null}
-          
-          <HStack mt="6" justifyContent="center">
-          <Button
-            mt="2"
-            colorScheme="primary"
-            onPress={handleSubmit(onSubmit)}
-          >
-            Add Child
-          </Button>
-          </HStack>
-        </VStack>
-    );
-  };
+ 
 
   return (
     <Box flex={1} safeAreaTop>
@@ -165,16 +81,14 @@ function ManageChildren({ }: Props) {
                 </Button>
               ))
             : null}
-          {!userLoading ? (
+
             <Button
               variant="ghost"
               w="100%"
               colorScheme="secondary"
               alignItems="center"
               justifyContent="flex-start"
-              onPress={() => {
-                setShowAddChildForm(true);
-              }}
+              onPress={toggleAddChildForm}
             >
               <Row alignItems="center" w="100%">
                 <Box>
@@ -187,8 +101,13 @@ function ManageChildren({ }: Props) {
                 </Box>
               </Row>
             </Button>
-          ) : null}
-          {renderAddChildForm()}
+          
+
+          {showAddChildForm && (
+            <VStack space={3} mt="5">
+              <AddChild onSubmitSuccess={toggleAddChildForm}/>
+            </VStack>
+          )}
         </Column>
       </Row>
       {/* <EditChildSheet
